+++ board/demo/configuration/rootfs/mnt/dnsmasq.conf
@@ -1,0 +1,11 @@
+log-facility=-
+no-daemon
+
+enable-tftp
+tftp-root=/mnt
+
+interface=eth0
+dhcp-option=66,"10.0.0.89"
+dhcp-option=67,"config.cfg"
+
+dhcp-range=tag:eth0,10.0.0.1,10.0.0.40,1h
