--- board/demo/usbboot/rootfs/etc/usbboot.json
@@ -1,6 +1,0 @@
-{
-    "vid": 2,
-
-    "west": [ "ethX5", "ethX6" ],
-    "east": [ "ethX7", "ethX8" ]
-}
