+++ configs/demo_configuration_defconfig
@@ -1,3 +1,4 @@
+BR2_x86_64=n
 BR2_powerpc=y
 BR2_powerpc_e5500=y
 BR2_ENABLE_DEBUG=y
@@ -9,10 +10,10 @@ BR2_TOOLCHAIN_EXTERNAL_GCC_7=y
 BR2_TOOLCHAIN_EXTERNAL_HEADERS_4_15=y
 BR2_TOOLCHAIN_EXTERNAL_CUSTOM_GLIBC=y
 BR2_TOOLCHAIN_EXTERNAL_GDB_SERVER_COPY=y
-BR2_TARGET_GENERIC_HOSTNAME="usbboot"
-BR2_TARGET_GENERIC_ISSUE="USBboot Demo"
-BR2_ROOTFS_OVERLAY="$(BR2_EXTERNAL_DEMO_PATH)/board/demo/usbboot/rootfs"
-BR2_ROOTFS_POST_BUILD_SCRIPT="$(BR2_EXTERNAL_DEMO_PATH)/board/demo/usbboot/post-build.sh"
+BR2_TARGET_GENERIC_HOSTNAME="configuration"
+BR2_TARGET_GENERIC_ISSUE="Configuration Demo"
+BR2_ROOTFS_OVERLAY="$(BR2_EXTERNAL_DEMO_PATH)/board/demo/configuration/rootfs"
+BR2_ROOTFS_POST_BUILD_SCRIPT="$(BR2_EXTERNAL_DEMO_PATH)/board/demo/configuration/post-build.sh"
 BR2_PACKAGE_DNSMASQ=y
 BR2_PACKAGE_DROPBEAR=y
 BR2_PACKAGE_LLDPD=y
@@ -27,4 +28,4 @@ BR2_PACKAGE_SOCAT=y
 BR2_PACKAGE_TCPDUMP=y
 BR2_TARGET_ROOTFS_SQUASHFS=y
 # BR2_TARGET_ROOTFS_TAR is not set
-BR2_PACKAGE_USBBOOT=y
+BR2_PACKAGE_CONFIGURATION=y
