+++ package/configuration/S80configuration
@@ -1,0 +1,25 @@
+#!/bin/sh
+
+#[ -f /etc/configuration/config.json -a -f /etc/configuration/id ] || exit 0
+
+case "$1" in
+	start)
+		printf "Starting configuration: "
+		start-stop-daemon -S -x /usr/sbin/configuration
+		[ $? = 0 ] && echo "OK" || echo "FAIL"
+		;;
+	stop)
+		printf "Stopping configuration: "
+		start-stop-daemon -K -q -x /usr/sbin/configuration
+		[ $? = 0 ] && echo "OK" || echo "FAIL"
+		;;
+	restart|reload)
+		$0 stop
+		$0 start
+		;;
+	*)
+		echo "Usage: $0 {start|stop|restart}"
+		exit 1
+esac
+
+exit 0
