--- package/usbboot/S80usbboot
@@ -1,25 +1,0 @@
-#!/bin/sh
-
-#[ -f /etc/usbboot/config.json -a -f /etc/usbboot/id ] || exit 0
-
-case "$1" in
-	start)
-		printf "Starting usbboot: "
-		start-stop-daemon -S -x /usr/sbin/usbboot
-		[ $? = 0 ] && echo "OK" || echo "FAIL"
-		;;
-	stop)
-		printf "Stopping usbboot: "
-		start-stop-daemon -K -q -x /usr/sbin/usbboot
-		[ $? = 0 ] && echo "OK" || echo "FAIL"
-		;;
-	restart|reload)
-		$0 stop
-		$0 start
-		;;
-	*)
-		echo "Usage: $0 {start|stop|restart}"
-		exit 1
-esac
-
-exit 0
