+++ src/configuration/LICENSE
