+++ src/configuration/README.md
@@ -1,0 +1,106 @@
+Download WeOS cfg from configuration application
+=====================================================
+
+Introduktion
+-------------
+
+The `configuration` application reads a non native configuration format
+from usb and converts it to a OS nativ format. The host will download
+the file using TFTP and use it as a new running config:
+
+```
+Host            #  Configuration Application
+                #  .-------------.  .-------------.
+                #  | DHCP-server |  | TFTP-server |
+                #  '------+------'  '------+------'
+.------------.  #         '------.  .------'
+|    cfg     |  #              .-+--+-.
+| inet: dhcp |  #              | eth0 |
+'------+-----'  #              '--+---'
+       |        #                 |
+       '--------------------------'
+```
+
+
+A simple text file containing a configuration in non native OS format is read and
+converted to WeoS JSON format by the configuration application.
+The application container will be connected to the host via a VETH pair, cfg/eth0.
+The host will download the converted configuration file using TFTP and use it as
+a new running-config.
+
+Configuration
+-------------
+
+### Prepare External Media
+
+It is important that the USB media is formatted as ext2, ext3 or ext4. We also
+need to create four files on the USB media:
+
+- `dnsmasq.conf`: Configuration for [Dnsmasq][dnsmasq], which will serve as
+  both DHCP server and TFTP server in the container.
+- `cfg-provider.sh`: A script that will serve as an entry point for the
+  container. Any modification could be instrumented from here. Responsible
+  for starting Dnsmasq.
+- `dynamic-config.cfg`: The "dynamic" config that Dnsmasq will serve. Note
+that you can give this file whatever name you want. It needs to be consistent
+with what is set as DHCP option 67 in Dnsmasq though.
+- `config/net-config.cfg`: The bootstrap config that will carry the
+  configuration for our application container.
+
+Prepare the USB media by creating the first two files:
+
+**dnsmasq.conf**
+```
+log-facility=-
+no-daemon
+
+enable-tftp
+tftp-root=/mnt
+
+interface=eth0
+dhcp-option=66,"10.0.0.89"
+dhcp-option=67,"dynamic-config.cfg"
+
+dhcp-range=tag:eth0,10.0.0.1,10.0.0.40,1h
+```
+
+**cfg-provider.sh**
+```
+#!/bin/sh
+
+# Set up the default interface
+ip link set dev eth0 up
+ip addr add 10.0.0.89/24 dev eth0
+
+# Here is were we could start a program that would
+# modify /mnt/dynamic-config.cfg, or even generate it from scratch.
+
+# Start our DHCP/TFTP server Dnsmasq
+exec dnsmasq -C /etc/dnsmasq.conf
+```
+
+**Note** `cfg-provider.sh` needs to have the executable bit set.
+
+### config-to-be-converted.txt
+Simple configuration file in texfile format containing a hostname.
+I.e.
+hostname hostname-from-container
+
+### dynamic-config.cfg
+
+In our example `dynamic-config.cfg` is rather static and is generated by the application by
+converting the config in config-to-be-converted to WeOS JSON format.
+
+The file should look like this after the conversion.
+
+```
+{
+  "system": {
+    "hostname": "hostname-from-container"
+  }
+}
+```
+
+After reboot the container application starts up, the host fetches the new
+running-config from the container, and that config is applied, resulting in
+the `hostname` changing to "hostname-from-container".
