+++ src/configuration/util.h
